añadida la página de inicio y sus elementos
- Creada la carpeta "videos" en "../resources/videos/" para el video de fondo de la home.
- Renombrada la carpeta "model" a "models".
- Creado el archivo "TypingAnimation.js" en una nueva carpeta en "..models/home/" para ejecutar el script de la home.
- Creado el archivo que contiene los estilos de la home (Home.css).
- Actualizado "Home.jsx" para que muestre la nueva home.
