+++ src/pages/Home.jsx
@@ -1,11 +1,25 @@
-import React from 'react';
+import React from "react";
+import "../stylesheets/Home.css";
+import TypingAnimation from "../models/home/TypingAnimation";
 
-function Home() {
+const Home = () => {
+  const messageText = "───▐▀▄──────▄▀▌───▄▄▄▄▄▄▄ ───▌▒▒▀▄▄▄▄▀▒▒▐▄▀▀▒██▒██▒▀▀▄ ──▐▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▀▄ ──▌▒▒▒▒▒▒▒▒▒▒▒▒▒▄▒▒▒▒▒▒▒▒▒▒▒▒▒▀▄ ▀█▒▒█▌▒▒█▒▒▐█▒▒▀▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▌ ▀▌▒▒▒▒▒▀▒▀▒▒▒▒▒▀▀▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▐_▄▄ ▐▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▄█▒█ ▐▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒█▀ ──▐▄▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▄▌  [CAMBIAR POR TEXTO DESPUÉS 🤙]";
+  const typingSpeed = 100;
+
+  const displayedMessage = TypingAnimation(messageText, typingSpeed);
+
   return (
-    <div>
-      <h1>Bienvenida a la sección Home</h1>
+    <div className="home-bg">
+      <video autoPlay muted loop id="home-video">
+        <source src="../resources/videos/fondo.mp4" type="video/mp4" />
+      </video>
+      <div className="text-container">
+        <div className="message">
+          {displayedMessage}
+        </div>
+      </div>
     </div>
   );
-}
+};
 
 export default Home;
