+++ src/stylesheets/App.css
@@ -6,7 +6,7 @@
 .main-content {
   font-family: Pangolin;
   margin-left: 100px;
-  padding: 25px;
+  padding: 0px;
   transition: margin-left 400ms ease;
 }
